app/api/deps.py
from datetime import datetime
from typing import Generator

from fastapi import Depends, HTTPException, status
from fastapi.security import OAuth2PasswordBearer
from jose import jwt
from pydantic import ValidationError
from sqlalchemy import and_
from sqlalchemy.orm import Session

from app.models import Users, Tokens
from app import schemas
from app.core.config import API_PATH
from app.core.config import SECRET_KEY, ALGORITHM
from app.db.session import SessionLocal

reusable_oauth2 = OAuth2PasswordBearer(
    tokenUrl=f"{API_PATH}/login/access-token"
)


def get_db() -> Generator:
    try:
        db = SessionLocal()
        yield db
    finally:
        db.close()


def get_current_user(db: Session = Depends(get_db), token: str = Depends(reusable_oauth2)) -> Users:
    try:
        payload = jwt.decode(
            token, SECRET_KEY, algorithms=[ALGORITHM]
        )
        token_data = schemas.TokenPayload(**payload)
    except(jwt.JWTError, ValidationError):
        raise HTTPException(
            status_code=status.HTTP_403_FORBIDDEN,
            detail='Could not validate credentials'
        )

    user = db.query(Users).filter(Users.id == token_data.sub).first()

    if not user:
        raise HTTPException(status_code=404, detail='User not found')
    return user.__dict__
