requirements.txt
aiofiles==0.7.0
alembic==1.6.4
bcrypt==3.2.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
ecdsa==0.14.1
fastapi==0.65.1
greenlet==1.1.0
h11==0.12.0
idna==2.10
Jinja2==3.0.1
Mako==1.1.4
MarkupSafe==2.0.1
mysqlclient==2.0.3
orjson==3.5.2
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.20
pydantic==1.8.2
python-dateutil==2.8.1
python-editor==1.0.4
python-jose==3.2.0
python-multipart==0.0.5
requests==2.25.1
rsa==4.7.2
six==1.16.0
SQLAlchemy==1.4.15
starlette==0.14.2
typing-extensions==3.10.0.0
urllib3==1.26.4
uvicorn==0.13.4
